+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Task like `npm run lint:fix` get turned into `npm_run_lint-fix` (so the colons don't screw up Gradle)
+- When `npm run` commands fail, they dump their console output as a Gradle error.
 
 ## [1.1.0] - 2024-08-04
 ### Added